- Rename package structure for LlmPromptContextualizer.kt to match new directory structure. - Update import paths in DefaultLlmProvider.kt to reflect new package structure.
